Fixed back navigating to Nav Activity, added Player title marquee-ing
The Nav Activity better manages its service listener to show the correct track info and selection when navigating back from the player. The player's title text view now marquees if it's too long. The player service now re-shuffles the track list if it reaches the end of the list while in shuffle mode.