Set up Playlist interaction, fixed tracks retrieved in PlaylistTask
The Playlist fragment now switches between a list of playlists and a list of tracks from the selected playlist.
When trying to play tracks from the playlists, I was getting errors retrieving the stream url. I had to make sure that each track in the playlists has a valid SoundCloud object.